Support other endpoints and generalize settings (#24)
This pull request makes it so the endpoint and general settings used for
the model are easier to customize. Addresses
https://github.com/C-Loftus/talon-ai-tools/issues/11

Also made some the documentation in the readme better and more
reflective of the current state of the repository.

Only a thing I'm not entirely sure on as if we want a different default
system prompt. For OpenAI, you do not have to specify the system prompt
and then there's some sort of implicit system prompt used in the
background. I don't think there's a way to know what that is. But for
some models like llamafiles, you need to explicitly state that so i
picked a system prompt that generally seems fine.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>