Codeium functionality (#46)
Adds very basic functionality to use [Codium](https://codeium.com/).
This is a free alternative to copilot. It is less feature rich but still
generally good.

@pokey curious if you have any ideas or opinions. (i.e. is this worth
adding or just bloat, any other offline copilot tools that might be
worth adding, etc.) Wanted to check since you made the copilot commands
and Codium is very similar. Wasn't sure to what extent you still use
them, or if you think this is useful.

I was waiting to add the cursorless support for codeium since I am not
sure how useful this is

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Joshua Aresty <jaresty@gmail.com>