Custom prompts and easier customization (#53)
Instead of using a raw talon list, we now use a capture
`user.modelPrompt` for all the model commands. The capture is formed via
the set union of static prompts and custom prompts. Custom prompts are
empty by default. There is an example that can be copied into the user
directory from the GPT folder.

The user can create a custom prompt talon list file outside this repo in
their user directory and it will automatically be the most specific
(i.e. you don't need to do weird context matching on the hostname or
stuff that is semantically meaningless)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>