Revert cursorless gpt (#83)
The long term goal for this repo is to have one single grammar for the
entire pipeline. Currently it is difficult to fit cursorless in that due
to the fact that it increases DFA compilation times for Talon and it is
occasionally too verbose/increaes cognitive load. However, it is still
quite useful, so adding it back for the time being until there is a
better plan for how to put this in the pipeline

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>