Use a better state class; more general refactor work (#127)
Changes the code to use a central state class instead of global
variables. This makes the code cleaner and easier to extend.

Any methods dealing with state like threading or contact we want to put
within this class to separate from pure functions or talon user facing
code.

Cleaned up a few other things. gpt.talon included "normal" as the
default but "normal" was not being checked anywhere. So just used ""
instead.

Don't believe this causes any regressions and is just a refactor.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>