Refactored to use typed dicts. Removed stale code (#144)
Kept same behavior the same as far as I can tell.

- Removed all dict with `any` as a key or value. Try to use `GPTMessage`
or `GPTMessageItem` as the type as much as possible. Let me know if
there is a name you prefer instead of these. I wasn't sure what to name
them.
- Got rid of old tts and image description code that I don't think is
relevant anymore. These don't follow the grammar and aren't being used
since they were broken but no one reported an issue
-  No pylance errors are being thrown under `standard` type checking
- Got rid of the a11y editor context logic in the system prompt. I will
probably add something to make that more explicit. I don't like that it
would be always passed in if the user has screen reader support enabled
for axkit
- Use the word `{user.model} **start** thread:` since it is one syllable
instead of `enable` which is 3 syllables.
- Changed `contextual_user_context` to be `gpt_additional_user_context`
since it is a globally exposed talon python function, we want to prefix
it with `gpt`. I personally find `additional_user_context` to be more
clear than `contextual_user_context` which strikes me as a bit
confusing.

I am using `TypeDict` so we keep all the methods within the standard
`dict` class. Was considering creating a super class for the message and
inheriting from it but I think this abstraction does not actually
benefit and adds boilerplate.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>