Better model window for threading (#147)
- renamed confirmation gui functions so they have proper naming style
- Made the text that we operate on optional as a param. We were
previously passing empty user messages with "" as text. If it is none,
we want to ignore this and not add it.
- Added a window for the thread. If you have the thread started and the
window is open, it will auto update the window with the new thread data.
- If you don't specify a model destination when the window is open, it
will always add to the window
- added a model toggle window function to see the thread (I am ok with
getting rid of this, not sure if it is proper if we already have model
pass thread to window. The issue with the latter is that it is difficult
to know when it is coming from the thread in order to format it properly

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>